fix: 게시글 쿼리 및 뮤테이트 분리
